requirements.txt
django
numpy
matplotlib
djangorestframework